added-payment

+++ utils/utils.ts
@@ -18,7 +18,7 @@ export const formatTimestamp = (
 
   const date = new Date(timestamp);
   const day = date.getDate().toString().padStart(2, "0");
-  const month = date.toLocaleString("default", { month: "long" });
+  const month = date.toLocaleString("default", { month: "short" });
   const year = date.getFullYear();
   let hours = date.getHours();
   const minutes = date.getMinutes().toString().padStart(2, "0");
